agora sim autor

+++ src/main/java/com/biblioteca/biblioteca/services/AutorService.java
@@ -1,5 +1,6 @@
 package com.biblioteca.biblioteca.services;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,12 +9,17 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.biblioteca.biblioteca.domain.Autor;
+import com.biblioteca.biblioteca.domain.LivroAutor;
+import com.biblioteca.biblioteca.exceptions.LivrosAutoresException;
 import com.biblioteca.biblioteca.repository.AutorRepository;
 
 @Service
 public class AutorService {
     @Autowired
     private AutorRepository autorRepository;
+
+    @Autowired
+    private LivroAutorService livroAutorService;
 
     public Page<Autor> getAll(Pageable pageable){
         return this.autorRepository.findAll(pageable);
@@ -28,7 +34,13 @@ public class AutorService {
     }
 
     public void deleteAuth(Autor autor){
-        
-        this.autorRepository.delete(autor);
+        List<LivroAutor> livrosAutoresAuth = this.livroAutorService.livrosAutoresAuth(autor);
+        if (livrosAutoresAuth.isEmpty()){
+            this.autorRepository.delete(autor);
+        }
+        else{
+            throw new LivrosAutoresException();
+        }
+
     }
 }
